fix: remove side effects from project selector
