+++ src/utils.ts
@@ -1,3 +1,5 @@
+import type { Project } from "./storage/db";
+
 // Can these be reordered without breaking older versions?
 // seems like no because they are bound to ints
 export enum MessageRequest {
@@ -15,4 +17,18 @@ export enum MessageRequest {
   SAVE_WINDOW_TO_ACTIVE_PROJECT,
   PING,
   RENAME_PROJECT
+}
+
+export async function getOrderedProjects(): Promise<Project[]> {
+  let projects = await chrome.runtime.sendMessage({ type: MessageRequest.GET_ALL_PROJECTS });
+  const activeProject = await chrome.runtime.sendMessage({ type: MessageRequest.GET_ACTIVE_PROJECT });
+
+
+  for (let i = 0; i < projects.length; i++) {
+    if (projects[i].name === activeProject.name) {
+      projects.splice(i, 1);
+    }
+  }
+  projects.unshift(activeProject);
+  return projects
 }
